Switched from nint() to int() in Grid_DepositBaryons.C. This fixes a vexing problem with the gravitational potential arising from the baryon deposition for l=2 grids that are not aligned with the root grid.
--HG--
branch : week-of-code

+++ src/enzo/Grid_DepositBaryons.C
@@ -105,14 +105,14 @@ int grid::DepositBaryons(grid *TargetGrid, FLOAT DepositTime)
        grid.  It must not extended beyond the active region of TargetGrid
        (if we are depositing in self). */
  
-    GridOffset[dim] = nint((GridLeftEdge[dim] -
+    GridOffset[dim] = int((GridLeftEdge[dim] -
 			    TargetGrid->GravitatingMassFieldLeftEdge[dim])/
 
 			   TargetGrid->GravitatingMassFieldCellSize) - 1; 
 
     if (TargetGrid == this)
       GridOffset[dim] = max(GridOffset[dim],
-	nint((TargetGrid->GridLeftEdge[dim] -
+	int((TargetGrid->GridLeftEdge[dim] -
 	      TargetGrid->GravitatingMassFieldLeftEdge[dim])/
 	     TargetGrid->GravitatingMassFieldCellSize) );
